requirements.txt
broadlink==0.19.0
cffi==1.17.0
cryptography==43.0.0
pycparser==2.22
pygame==2.6.0
python-dotenv==1.0.1
